chore: migrate to uvu & maven-like dir layout (#1)
