--- component.json
@@ -1,24 +1,0 @@
-{
-  "name": "toposort",
-  "author": "Marcel Klehr <mklehr@gmx.net>",
-  "repo": "marcelklehr/toposort",
-  "description": "Topological sort of directed acyclic graphs (like dependency lists)",
-  "version": "0.2.10",
-  "keywords": [
-    "topological",
-    "sort",
-    "sorting",
-    "graphs",
-    "graph",
-    "dependency",
-    "list",
-    "dependencies",
-    "acyclic"
-  ],
-  "dependencies": {},
-  "development": {},
-  "license": "MIT",
-  "scripts": [
-    "index.js"
-  ]
-}
