+++ package.json
@@ -1,17 +1,23 @@
 {
-  "name": "toposort",
-  "version": "2.0.2",
+  "name": "@qiwi/toposort",
+  "version": "0.0.0",
   "description": "Topological sort of directed ascyclic graphs (like dependecy lists)",
-  "main": "index.js",
+  "main": "src/main/js/index.js",
   "scripts": {
-    "test": "node test.js"
+    "test": "node src/test/js/index.js"
   },
+  "files": [
+    "src/main",
+    "README.md",
+    "package.json"
+  ],
+  "type": "module",
   "repository": {
     "type": "git",
-    "url": "https://github.com/marcelklehr/toposort.git"
+    "url": "https://github.com/qiwi-forks/toposort.git"
   },
   "devDependencies": {
-    "vows": "0.7.x"
+    "uvu": "^0.5.6"
   },
   "keywords": [
     "topological",
